波次内补货--工作优先级的修改和按目标库位容量逐步释放的修改

+++ warehouse-operation-api/src/main/java/com/baozun/scm/primservice/whoperation/manager/odo/wave/proxy/CreateWorkInWaveManagerProxy.java
@@ -232,7 +232,7 @@ public interface CreateWorkInWaveManagerProxy extends BaseManager {
      * @param List<WhOdoOutBoundBox>
      * @return
      */
-    public Double locationReplenishmentCalculation(WhSkuInventoryAllocatedCommand siaCommand, Long ouId);
+    public Long locationReplenishmentCalculation(WhSkuInventoryAllocatedCommand siaCommand, Long ouId);
     
     /**
      * [业务方法] 创建补货工作-更新波次头信息
